fix: avoid circular reference
